requirements.txt
django>=1.4.2                                               # django framework
psycopg2                                                    # database adaptor
django-reversion>=1.6.4                                     # versioning
South>=0.7.5                                                # database migrations
django-robots                                               # robots.txt
pytz                                                        # timezone support
django-secure                                               # security
django-discover-runner                                      # better test runner
django-tables2                                              # better table handling
django-watson                                               # searching
django-mptt                                                 # mptt for classification app
